refactor: RequestHandler 클래스 제거
